+++ packages/server/.eslintrc.cjs
@@ -18,6 +18,7 @@ module.exports = {
   rules: {
     '@typescript-eslint/ban-ts-comment': 1,
     '@typescript-eslint/no-unused-vars': 0,
+    'max-classes-per-file': 0,
     'no-console': 1,
   }
 }
